W-15188084: Encapsulate org.mule.runtime.core.privileged.util
* Prework: Utils used by api-gateway
