+++ core-tests/src/test/java/org/mule/runtime/core/internal/util/version/VersionRangeTestCase.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.core.internal.util;
+package org.mule.runtime.core.internal.util.version;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
