+++ core/src/main/java/org/mule/runtime/core/internal/util/OneTimeWarning.java
@@ -14,69 +14,13 @@ import org.slf4j.Logger;
  * warn that they should not be used anymore.
  *
  * @since 3.6.0
+ * @deprecated Use {@link org.mule.runtime.core.internal.util.log.OneTimeWarning} instead
  */
-public class OneTimeWarning {
-
-  private final Logger logger;
-  private final String message;
-  private boolean warned = false;
-  private Delegate delegate = new FirstTimeDelegate();
+@Deprecated
+public class OneTimeWarning extends org.mule.runtime.core.internal.util.log.OneTimeWarning {
 
   public OneTimeWarning(Logger logger, String message) {
-    this.logger = logger;
-    this.message = message;
+    super(logger, message);
   }
 
-  /**
-   * Fire one time warning.
-   */
-  public void warn() {
-    delegate.warn();
-  }
-
-  /**
-   * Fire one time warning using arguments passing arguments to logger
-   *
-   * @param args arguments to pass to logger along with message defined when this instance was created.
-   */
-  public void warn(Object... args) {
-    delegate.warn(args);
-  }
-
-  private interface Delegate {
-
-    void warn();
-
-    void warn(Object... args);
-  }
-
-  private class FirstTimeDelegate implements Delegate {
-
-    @Override
-    public synchronized void warn() {
-      warn(null);
-    }
-
-    @Override
-    public synchronized void warn(Object... args) {
-      if (warned == false) {
-        if (args != null) {
-          logger.warn(message, args);
-        } else {
-          logger.warn(message);
-        }
-        warned = true;
-        delegate = new NoOpDelegate();
-      }
-    }
-  }
-
-  private class NoOpDelegate implements Delegate {
-
-    @Override
-    public void warn() {}
-
-    @Override
-    public void warn(Object... args) {}
-  }
 }
