+++ modules/extensions-support/src/main/java/org/mule/runtime/module/extension/internal/loader/AbstractExtensionModelLoader.java
@@ -10,8 +10,8 @@ import static org.mule.runtime.api.util.JavaConstants.JAVA_VERSION_8;
 import static org.mule.runtime.api.util.MuleSystemProperties.DISABLE_SDK_IGNORE_COMPONENT;
 import static org.mule.runtime.api.util.MuleSystemProperties.ENABLE_SDK_POLLING_SOURCE_LIMIT;
 import static org.mule.runtime.api.util.Preconditions.checkState;
-import static org.mule.runtime.core.internal.util.JdkVersionUtils.getJdkVersion;
-import static org.mule.runtime.core.internal.util.JdkVersionUtils.isJava8;
+import static org.mule.runtime.core.internal.util.version.JdkVersionUtils.getJdkVersion;
+import static org.mule.runtime.core.internal.util.version.JdkVersionUtils.isJava8;
 import static org.mule.runtime.extension.api.ExtensionConstants.VERSION_PROPERTY_NAME;
 import static org.mule.runtime.module.extension.internal.ExtensionProperties.DISABLE_COMPONENT_IGNORE;
 import static org.mule.runtime.module.extension.internal.ExtensionProperties.ENABLE_POLLING_SOURCE_LIMIT_PARAMETER;
@@ -22,8 +22,7 @@ import static java.lang.System.getProperty;
 import static java.util.Arrays.asList;
 import static java.util.Collections.unmodifiableList;
 
-import org.mule.runtime.core.internal.util.JdkVersionUtils;
-import org.mule.runtime.core.internal.util.JdkVersionUtils.JdkVersion;
+import org.mule.runtime.core.internal.util.version.JdkVersionUtils.JdkVersion;
 import org.mule.runtime.extension.api.loader.ExtensionLoadingContext;
 import org.mule.runtime.extension.api.loader.ExtensionModelLoader;
 import org.mule.runtime.extension.api.loader.ExtensionModelValidator;
@@ -36,7 +35,6 @@ import org.mule.runtime.module.extension.internal.loader.validator.JavaConnectio
 import org.mule.runtime.module.extension.internal.loader.validator.ParameterPluralNameModelValidator;
 import org.mule.runtime.module.extension.internal.runtime.operation.IllegalSourceException;
 
-import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
