+++ modules/extensions-support/src/test/java/org/mule/runtime/module/extension/internal/manager/jdk/NullExtensionJdkValidatorTestCase.java
@@ -9,7 +9,7 @@ package org.mule.runtime.module.extension.internal.manager.jdk;
 import static org.mule.test.allure.AllureConstants.Sdk.SDK;
 import static org.mule.test.allure.AllureConstants.Sdk.SupportedJavaVersions.ENFORCE_EXTENSION_JAVA_VERSION;
 
-import org.mule.runtime.core.internal.util.JdkVersionUtils.JdkVersion;
+import org.mule.runtime.core.internal.util.version.JdkVersionUtils.JdkVersion;
 import org.mule.tck.size.SmallTest;
 
 import io.qameta.allure.Feature;
